Add files via upload
In the context of chemical communication in insects, where pheromones play a vital role, the endocrine regulation of their production is a complex and little-studied area, particularly in fire ants. The researchers therefore turned to the PBAN peptide, known for its role in pheromone biosynthesis in Lepidoptera, with the aim of extending our knowledge of trail pheromones in fire ants. In this context, the protagonists are fire ants, the neuropeptide PBAN and its receptor R-PBAN, and trailing pheromones. This choice of study is based on the poorly developed knowledge of how fire ants follow each other when one of them finds a resource they need. What's more, these protagonists have already been discovered as activators of sex pheromone biosynthesis in moths. The results obtained demonstrated that PBAN effectively regulates pheromone biosynthesis in these insects, opening up new avenues of research into the endocrine regulation of pheromones in different insect species. The aim is to see if  PBAN interacts with its receptor, thereby influencing the production of drag pheromones in these ants. To achieve this goal, the researchers used fluorescence and gene expression inhibition strategies targeting the PBAN gene or its receptor in fire ants.